Fix windows build

+++ engine/data_reader.cpp
@@ -16,7 +16,7 @@ void DataReader::Reset(std::vector<Ui8> &&in_data) {
 
 Ui64 DataReader::Read(void *dst, Ui64 amount) {
   Ui64 to_read = std::min(amount, (Ui64)(end - p));
-  memcpy(dst, p, to_read);
+  memcpy(dst, p, (size_t)to_read);
   return to_read;
 }
